Added namespace.
Added the namespace LearningCurve\BeansVisualHookGuide and
the __NAMESPACE__ placeholder for each of the callbacks.

Why?  The namespace packages up your plugin, giving it a
highly readable and specific naming convention.  The result is:

1. It's more readable.
2. Less chance of a naming conflict with another plugin.

It sets us up to remove the prefix from each of the functions.

Why would we want to remove the prefix?  The prefix makes
the name less readable.  A function's name should describe
the expected behavior of the function.  Having a prefix means
you need to skip it over it when you are reading the function
to see what it does.