Removed prefix from functions.
The prefix is not needed anymore as we are using namespacing.

As I noted in the last commit, the prefix makes the code less
readable. Why?

A function should be named by its expected behavior, starting
with a verb.  A function runs a task.  It's the worker.  To
quickly comprehend the code, it is important that we name
functions to _tell us_ what it will do when we call it.

With a prefix, you have to sort out what part of the name is
for the prefix and the remaining part for the expected behavior.
That is distracting and less readable.