Refactored the $_GET checkers.
The $_GET checkers had a bug as isset( $_GET['whatever'] ) returns
true or false.  It does not return if it is set to "show."

Refactored each of the $_GET checkers, adding functions in the
plugin.php file.  Why? To keep the code DRY as that we don't
have to repeat:

1. isset()
2. === or !== "show"

Instead, we use one function that handles it for all instances.