Refactored "CSS on the Fly".
Previously, this functionality required a global variable (yuck).
To eliminate the global, we need a different strategy.

This functionality is a perfect candidate for a Singleton, as
we want to collect selected markup IDs and abstract away the
handling of the enqueuing and storage processes.

Bam, new class `Css_On_The_Fly`.