Refactored Admin Bar.
The previous design of the menu adding system for the Admin bar
was confusing and not overly readable.  It also required global
variables to hold state between disconnected functions.

The new design uses OOP.  Advantages are:

1. Encapsulation to hide away complexity.
2. Eliminates global variables.
3. Hold state and configurations inside of the object itself.

For the configurable parameters, this commit moves those to a
configuration file.  This file is loaded into the object upon
instantiation.