Changed asset version to file modification time.
When enqueueing, this commit uses the file's modification
timestamp instead of a hard-coded plugin version number.

Why?

Each time a stylesheet or script code changes, its file's
modification time changes.  With the previous design, we
have to remember to uptick the plugin's version number in
order for the new file to be served to the browser.  We
might forget or be in the middle of development.

With this new design, we leverage the filesystem.  When the
file changes, change the version number.  Then the browser
is forced to request the new file instead of using the
previously cached on.  That means, the browser will go get
the updated file from the server and we see the new stuff.

This commit also removes the unused _get_plugin_version()
function.